// src/generated/types/StudentWeeklyScheduleDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type StudentWeeklyScheduleDto = {
  /**
   * @description Course ID
   * @type string
   */
  courseId: string
  /**
   * @description Course name
   * @type string
   */
  courseName: string
  /**
   * @description Course code like \"CS101\"
   * @type string
   */
  courseCode: string
  /**
   * @description Group name like \"Group A\"
   * @type string
   */
  groupName: string
  /**
   * @description Lab name and location
   * @type string
   */
  labName: string
  /**
   * @description Day of the week
   * @type string
   */
  weekDay: string
  /**
   * @description Start time in HH:MM format
   * @type string
   */
  startTime: string
  /**
   * @description End time in HH:MM format
   * @type string
   */
  endTime: string
  /**
   * @description Teaching assistant names
   * @type array
   */
  teachingAssistants: string[]
}